+++ crates/scheduler/src/event.rs
@@ -24,6 +24,7 @@ use wasefire_logger as log;
 use crate::Scheduler;
 
 pub mod button;
+pub mod radio;
 pub mod timer;
 pub mod uart;
 pub mod usb;
@@ -39,6 +40,7 @@ pub struct InstId;
 #[derivative(Ord = "feature_allow_slow_enum")]
 pub enum Key<B: Board> {
     Button(button::Key<B>),
+    Radio(radio::Key),
     Timer(timer::Key<B>),
     Uart(uart::Key<B>),
     Usb(usb::Key),
@@ -62,6 +64,7 @@ impl<'a, B: Board> From<&'a Event<B>> for Key<B> {
     fn from(event: &'a Event<B>) -> Self {
         match event {
             Event::Button(event) => Key::Button(event.into()),
+            Event::Radio(event) => Key::Radio(event.into()),
             Event::Timer(event) => Key::Timer(event.into()),
             Event::Uart(event) => Key::Uart(event.into()),
             Event::Usb(event) => Key::Usb(event.into()),
@@ -106,6 +109,7 @@ pub fn process<B: Board>(scheduler: &mut Scheduler<B>, event: Event<B>) {
     let mut params = vec![*func, *data];
     match event {
         Event::Button(event) => button::process(event, &mut params),
+        Event::Radio(event) => radio::process(event),
         Event::Timer(_) => timer::process(),
         Event::Uart(_) => uart::process(),
         Event::Usb(event) => usb::process(event),