+++ crates/board/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Major
 
+- Add `Api::Radio` for radio APIs
+- Add `radio::Api::Ble` for Bluetooth Low Energy
 - Add `platform::Api::Update` (fix #47)
 - Add `Api::Platform` with `platform::Api::reboot()` function
 - Require the APIs to be `Send` and the top-level API to be `'static` too
@@ -93,4 +95,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 12 -->
+<!-- Increment to skip CHANGELOG.md test: 14 -->